+++ adapters/messages_adapter.py
