+++ adapters/network_adapter.py
